[ruby/rdoc] Fix sidebar scroll again and add missing footer back (https://github.com/ruby/rdoc/pull/1154)
* Add missing footers

In #1152 the footer partial was only added to the index.rhtml file.
This commit adds the footer partial to the other template files.

* Remove unnecessary middle divs in nav

* Simplify sidebar's overflow settings

Because sidebar needs to be scrollable, its overflow should default to auto.
Currently it's set to hidden and force individual elements to set overflow auto,
which overcomplicates things.

https://github.com/ruby/rdoc/commit/b8c2bcd8db