[rubygems/rubygems] Don't break if extra calls to File.writable? happen
In https://bugs.ruby-lang.org/issues/20693, I'd like to have Dir.tmpdir
call `File.writable?` instead of `Stat#writable?`. However, that causes
this test to break in bundler because it's using RSpec to stub
`File.writable?`.

We can fix this by allowing the real `File.writable?` to be called for
all files except the directory we're trying to stub.

https://github.com/rubygems/rubygems/commit/0fa6657293