Take into account pooled_slots when creating pages
[Bug #20762]

If we run with RGENGC_FORCE_MAJOR_GC, then all GC's will be major GC's.
This means that any pages added to pooled pages will be unable to be
allocated into until the next major GC. We should take these held slots
into account when deciding whether to allocate another page after a
major or not.

Co-Authored-By: Peter Zhu <peter@peterzhu.ca>