Windows: Change command line interface to UTF-8
... even when the Windows codepage isn't set to 65001/UTF-8.
A non-UTF-8 codepage is still the default on Windows-11.

The following inputs are changed to consistently work with UTF-8:
- script name
- include paths
- script input from stdin

They are currently a inconsistent mix of locale encoding and UTF-8, depending on the situation.

The adjusted tests don't trigger encoding issues when the codepage is 65001 (UTF-8), but they do on 850 or 437 or so.
Therefore it doesn't fail on Github Actions, which sets the codepage to UTF-8 per default.