[ruby/win32-registry] Move all tests to volatile environment and use per-test individual registry key
This allows to run all tests in parallel.
Usage of `File` methods is avoided.
Since all tests are executed in the volatile environment of the registry,
the dedicated 'create_volatile' test is removed now.
Also add some documentation to the test setup.

Downside of this use of the "Volatile Environment" is that we can not use or test
the `create` method with default options.
This is because within this path only keys with option `REG_OPTION_VOLATILE` are allowed.

https://github.com/ruby/win32-registry/commit/3c186ae7dd