[DOC] Use RIGHT SINGLE QUOTATION MARK
Ideally, RDoc would be expected to replace U+0027 APOSTROPHE with
U+2019 RIGHT SINGLE QUOTATION MARK, as well as the case the preceeding
word is not a class name; but a boundary of attributes is treated as a
word boundary currently.  Although I'm not sure if this is a bug of
RDoc, use that mark meanwhile.