Make StringScanner compactable, WB-protected and embedded
Modify strscan_mark and implement strscan_compact so that the reference
fields of StringScanner, i.e. "str" and "regex", can be updated in a
moving GC.

Also apply write barriers when updating the two fields.  That makes
StringScanner WB-protected.

Also make StringScanner an embedded RTypedData to reduce an unnecessary
malloc and free.