[rubygems/rubygems] Simplify tests that deal with lockfile using only "ruby" platform
Using an old bundler version was necessary at the beginning because we
checked the version of Bundler being run in order to enable the new
behavior. But we removed that a long time ago and now we only look at
whether the lockfile only includes "ruby" as a platform or not.

https://github.com/rubygems/rubygems/commit/8fc3d38ff3