[rubygems/rubygems] Remove override of worker jobs for `bundle install --local`
There seems to be no reason why the install should be serial for --local.
The packages are still installed in the right dependency order in this case, so that parallel install can be used.
This patch disables parallel install only in case of no_install_needed.

Also remove the `option` argument, which is effectifely not used.

https://github.com/rubygems/rubygems/commit/5da934ddb6