[ruby/digest] Fix loading of digest ext libraries with `--with-static-linked-ext`
`rb_ext_resolve_symbol` is not always available on some platforms
including WASI, and we don't need to use it when the extension is built
as a static library. This commit prefers to use `rb_digest_wrap_metadata`
directly when the extension is built as a static library to avoid the
unnecessary use of `rb_ext_resolve_symbol`.

https://github.com/ruby/digest/commit/f8ff014622