Fix CI
Bundler specs broke because of a change in Sinatra 4.1.0. In my opinion,
a change in an external test dependency should not break ruby-core's CI,
whose main purpose is to test ruby core code, not external dependencies.
Also, it makes PRs CI more useful, because it makes the rule "A PR is
red if and only if the PRs code breaks tests" closer to be followed.

This is how Bundler specs were run until
426434556f43e597da8ddfa33c659321971f1b83, but I'm not sure what the
motivation for that commit was.