allow malloc outside of GVL
We currently do not support calling ruby_xmalloc out of GVL.  OTOH
oftentimes people do `#define malloc ruby_xmalloc`.  Under GVL that
has no practical problem but without one it must be a nightmare.

Because whether a runtime context has GVL or not is not fixed when
the macro is expanded, why don't we actively check that property
on the fly and (instead of abnormal end of the process) return NULL.

Better safe than sorry I guess.