Add new `rb_enc_str_adopt` API
[Feature #20878]

A common use case when writing C extensions is to generate text or bytes
into a buffer an to return it back wrapped into a Ruby String.

Currently this requires to either work in a native buffer and then
copy into a Ruby String, or to pre-allocate a Ruby string and have
to use muchless performant APIs to grow the Ruby String when needed.

Instead it is much more convenient to ask Ruby to adopt the buffer
inside a Ruby String.

Such API impose a few constraints as noted in the documentation,
such as the provided buffer MUST be freeable by `ruby_xfree`.

The pointer is adopted as soon as the function is called, which
means if the String can't be allocated for any reason, the pointer
is immediately freed.