Improve `Socket.tcp_fast_fallback=` / `Socket.tcp_fast_fallback`
The changes in https://github.com/ruby/ruby/pull/12210 made it possible to globally control the default value of the fast_fallback option for `TCPSocket.new`.

In this update, the parts of the implementation that dynamically retrieved constant and variable names during that change were revisited, improving performance as follows:

(Case of 1000 executions of `TCPSocket.new` to the local host)

Rehearsal -----------------------------------------
before   0.031462   0.147946   0.179408 (  0.249279)
after    0.031164   0.146839   0.178003 (  0.346935)
-------------------------------- total: 0.178003sec

            user     system      total        real
before   0.027584   0.138712   0.166296 (  0.233356)
after    0.025953   0.127608   0.153561 (  0.237971)

Additionally, documentation for these methods has been added.