[ruby/rdoc] Enable cross reference in code (https://github.com/ruby/rdoc/pull/1240)
Some people like to mark up method names in MarkDown style block
quotes, like this: ruby/ruby#12333.
Currently, no links are created in the code in the RDoc, but such
words most likely refer to methods.
This PR makes a word a code cross-reference if the whole word can be
resolved as a reference.

https://github.com/ruby/rdoc/commit/7d7efb0709