Don't use xmalloc for newobj_i
newobj_i is a GC hook that runs when a new object is allocated. We should
not use xmalloc because it can potentially trigger GC and a lot of code
is unsafe to trigger a GC right after an object has been allocated because
they perform initialization for the object and assume that the GC does
not trigger before then.