[DOC] Resurrected YJIT section
under "Implementation improvements", similarly to past versions.

We also don't do categorization like "* New features" (which was in this
file) or "New features:" (which was on the release notes), so I used the
standard header format instead.

I applied the same change to release notes, copied the same thing back
to NEWS.md, and then removed the TL;DR section from it.